Create etch-a-sketch project
This is the first build of the etch-a-sketch. It lacks transparency or
color functionality, which is to be added. Additionally, the way the
grid is generated needs tweaking as the size of the entire grid does
not cleanly line up with the cells. This is apparent if you zoom in.